chore: Refactored Bootstraper into separate files
